ajuste categoria e inscricao

+++ SportsPromo.Dominio.Modelos/Categoria.cs
@@ -8,7 +8,7 @@ namespace SportsPromo.Dominio.Modelos
 {
     public class Categoria
     {
-        public int CategoriaId { get; set; }
+        public long CategoriaId { get; set; }
 
         public string CategoriaNome { get; set; }
 
@@ -16,7 +16,7 @@ namespace SportsPromo.Dominio.Modelos
 
         public int CategoriaIdadeMax { get; set; }
 
-        public int CategoriaGeneroId { get; set; }
+        public long CategoriaGeneroId { get; set; }
 
         public Genero Genero { get; set; }
     }
